feat(mcp): rename /mcp-oauth to /mcp/oauth, add OAUTH_USER_AGENTS
- Rename /mcp-oauth rewrite to /mcp/oauth in vercel.json
- Add OAUTH_USER_AGENTS env var to force OAuth on /mcp for matching user agents
- Move hasAuth/create401Response into mcp.ts, accept forceOAuth option
- Simplify mcp-oauth.ts to delegate to handleRequest with forceOAuth flag
- Add /mcp/oauth to pathname normalization
